Put team filter in MeetingsDashHeader
